refactor(link): 크롤 실패 잡 처리 로직 분리
문서 파싱, 자동 재시도 조건, 실패 잡 모델을 전용 구성요소로 분리하고 링크/실패 잡 길이 제약을 엔티티 상수로 모은다.
